Added rank_split_strategy and use blocks as default
This changes the default behaviour back to making blocks of the input
dataset and assign to ranks. This means that each rank likely gets it's
own season of data, instead of a mix of all seasons. This is beneficial
since it ensures that each batch is mixes, as gradients are averaged
across ranks.