Allow maelstrom-train to run based on an output config
This commit allows maelstrom-train to be run on config.yml that is
output by previous runs (for example to run testing after training has
been done separately). This deals with the fact that in a
config.yml file, the model is defined in the "Model" section, instead of
"models", which is used in the standard input config files.